# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "데이트 장소 만들기 위한 API",
        "title": "Swagger Date Path API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "paths": {
        "/direction/coordinates": {
            "get": {
                "description": "길찾기",
                "produces": [
                    "application/json"
                ],
                "summary": "길찾기",
                "parameters": [
                    {
                        "maxLength": 40,
                        "minLength": 3,
                        "type": "string",
                        "default": "37.389624,126.950745",
                        "description": "출발 지점의 위도,경도. default: 범계역",
                        "name": "origin",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 40,
                        "minLength": 3,
                        "type": "string",
                        "default": "37.394640,127.111135",
                        "description": "도착 지점의 위도,경도역. default: 판교역",
                        "name": "destination",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1614332700",
                        "description": "출발 시간. UTC 1970 년 1월 1일 0시 0분 0초 부터 경과한 초를 정수로 반환한 값. 현재 시각으로 찾으려면 'now'를 전달한다.",
                        "name": "departure_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/direction.Response"
                        }
                    }
                }
            }
        },
        "/local/search/keyword": {
            "post": {
                "description": "키워드 장소 검색",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "키워드 장소 검색",
                "parameters": [
                    {
                        "maxLength": 15,
                        "minLength": 1,
                        "type": "string",
                        "default": "판교역",
                        "description": "검색을 원하는 질의어",
                        "name": "query",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "중심 좌표의 X값 혹은 longitude. 특정 지역을 중심으로 검색하려고 할 경우 radius와 함께 사용 가능",
                        "name": "x",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "중심 좌표의 Y값 혹은 latitude. 특정 지역을 중심으로 검색하려고 할 경우 radius와 함께 사용 가능",
                        "name": "y",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "중심 좌표부터의 반경거리. 특정 지역을 중심으로 검색하려고 할 경우 중심좌표로 쓰일 x,y와 함께 사용. 단위 meter, 0~20000 사이의 값",
                        "name": "radius",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "direction.Response": {
            "type": "object",
            "properties": {
                "geocoded_waypoints": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "geocoder_status": {
                                "type": "string"
                            },
                            "place_id": {
                                "type": "string"
                            },
                            "types": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "routes": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "bounds": {
                                "type": "object",
                                "properties": {
                                    "northeast": {
                                        "type": "object",
                                        "properties": {
                                            "lat": {
                                                "type": "number"
                                            },
                                            "lng": {
                                                "type": "number"
                                            }
                                        }
                                    },
                                    "southwest": {
                                        "type": "object",
                                        "properties": {
                                            "lat": {
                                                "type": "number"
                                            },
                                            "lng": {
                                                "type": "number"
                                            }
                                        }
                                    }
                                }
                            },
                            "copyrights": {
                                "type": "string"
                            },
                            "legs": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "arrival_time": {
                                            "type": "object",
                                            "properties": {
                                                "text": {
                                                    "type": "string"
                                                },
                                                "time_zone": {
                                                    "type": "string"
                                                },
                                                "value": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "departure_time": {
                                            "type": "object",
                                            "properties": {
                                                "text": {
                                                    "type": "string"
                                                },
                                                "time_zone": {
                                                    "type": "string"
                                                },
                                                "value": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "distance": {
                                            "type": "object",
                                            "properties": {
                                                "text": {
                                                    "type": "string"
                                                },
                                                "value": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "duration": {
                                            "type": "object",
                                            "properties": {
                                                "text": {
                                                    "type": "string"
                                                },
                                                "value": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "end_address": {
                                            "type": "string"
                                        },
                                        "end_location": {
                                            "type": "object",
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            }
                                        },
                                        "start_address": {
                                            "type": "string"
                                        },
                                        "start_location": {
                                            "type": "object",
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            }
                                        },
                                        "steps": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "distance": {
                                                        "type": "object",
                                                        "properties": {
                                                            "text": {
                                                                "type": "string"
                                                            },
                                                            "value": {
                                                                "type": "integer"
                                                            }
                                                        }
                                                    },
                                                    "duration": {
                                                        "type": "object",
                                                        "properties": {
                                                            "text": {
                                                                "type": "string"
                                                            },
                                                            "value": {
                                                                "type": "integer"
                                                            }
                                                        }
                                                    },
                                                    "end_location": {
                                                        "type": "object",
                                                        "properties": {
                                                            "lat": {
                                                                "type": "number"
                                                            },
                                                            "lng": {
                                                                "type": "number"
                                                            }
                                                        }
                                                    },
                                                    "html_instructions": {
                                                        "type": "string"
                                                    },
                                                    "polyline": {
                                                        "type": "object",
                                                        "properties": {
                                                            "points": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    },
                                                    "start_location": {
                                                        "type": "object",
                                                        "properties": {
                                                            "lat": {
                                                                "type": "number"
                                                            },
                                                            "lng": {
                                                                "type": "number"
                                                            }
                                                        }
                                                    },
                                                    "steps": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object",
                                                            "properties": {
                                                                "distance": {
                                                                    "type": "object",
                                                                    "properties": {
                                                                        "text": {
                                                                            "type": "string"
                                                                        },
                                                                        "value": {
                                                                            "type": "integer"
                                                                        }
                                                                    }
                                                                },
                                                                "duration": {
                                                                    "type": "object",
                                                                    "properties": {
                                                                        "text": {
                                                                            "type": "string"
                                                                        },
                                                                        "value": {
                                                                            "type": "integer"
                                                                        }
                                                                    }
                                                                },
                                                                "end_location": {
                                                                    "type": "object",
                                                                    "properties": {
                                                                        "lat": {
                                                                            "type": "number"
                                                                        },
                                                                        "lng": {
                                                                            "type": "number"
                                                                        }
                                                                    }
                                                                },
                                                                "polyline": {
                                                                    "type": "object",
                                                                    "properties": {
                                                                        "points": {
                                                                            "type": "string"
                                                                        }
                                                                    }
                                                                },
                                                                "start_location": {
                                                                    "type": "object",
                                                                    "properties": {
                                                                        "lat": {
                                                                            "type": "number"
                                                                        },
                                                                        "lng": {
                                                                            "type": "number"
                                                                        }
                                                                    }
                                                                },
                                                                "travel_mode": {
                                                                    "type": "string"
                                                                }
                                                            }
                                                        }
                                                    },
                                                    "transit_details": {
                                                        "type": "object",
                                                        "properties": {
                                                            "arrival_stop": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "location": {
                                                                        "type": "object",
                                                                        "properties": {
                                                                            "lat": {
                                                                                "type": "number"
                                                                            },
                                                                            "lng": {
                                                                                "type": "number"
                                                                            }
                                                                        }
                                                                    },
                                                                    "name": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            },
                                                            "arrival_time": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "text": {
                                                                        "type": "string"
                                                                    },
                                                                    "time_zone": {
                                                                        "type": "string"
                                                                    },
                                                                    "value": {
                                                                        "type": "integer"
                                                                    }
                                                                }
                                                            },
                                                            "departure_stop": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "location": {
                                                                        "type": "object",
                                                                        "properties": {
                                                                            "lat": {
                                                                                "type": "number"
                                                                            },
                                                                            "lng": {
                                                                                "type": "number"
                                                                            }
                                                                        }
                                                                    },
                                                                    "name": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            },
                                                            "departure_time": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "text": {
                                                                        "type": "string"
                                                                    },
                                                                    "time_zone": {
                                                                        "type": "string"
                                                                    },
                                                                    "value": {
                                                                        "type": "integer"
                                                                    }
                                                                }
                                                            },
                                                            "headsign": {
                                                                "type": "string"
                                                            },
                                                            "headway": {
                                                                "type": "integer"
                                                            },
                                                            "line": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "agencies": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "type": "object",
                                                                            "properties": {
                                                                                "name": {
                                                                                    "type": "string"
                                                                                },
                                                                                "url": {
                                                                                    "type": "string"
                                                                                }
                                                                            }
                                                                        }
                                                                    },
                                                                    "color": {
                                                                        "type": "string"
                                                                    },
                                                                    "name": {
                                                                        "type": "string"
                                                                    },
                                                                    "short_name": {
                                                                        "type": "string"
                                                                    },
                                                                    "text_color": {
                                                                        "type": "string"
                                                                    },
                                                                    "vehicle": {
                                                                        "type": "object",
                                                                        "properties": {
                                                                            "icon": {
                                                                                "type": "string"
                                                                            },
                                                                            "name": {
                                                                                "type": "string"
                                                                            },
                                                                            "type": {
                                                                                "type": "string"
                                                                            }
                                                                        }
                                                                    }
                                                                }
                                                            },
                                                            "num_stops": {
                                                                "type": "integer"
                                                            }
                                                        }
                                                    },
                                                    "travel_mode": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        },
                                        "traffic_speed_entry": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "via_waypoint": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    }
                                }
                            },
                            "overview_polyline": {
                                "type": "object",
                                "properties": {
                                    "points": {
                                        "type": "string"
                                    }
                                }
                            },
                            "summary": {
                                "type": "string"
                            },
                            "warnings": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "waypoint_order": {
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}